chat

+++ src/main/java/com/websocket/chat/entity/ChatRoomMembersPK.java
@@ -1,14 +1,14 @@
 package com.websocket.chat.entity;
 
-import jakarta.persistence.Embeddable;
+import javax.persistence.Embeddable;
 
 import java.io.Serializable;
 
 @Embeddable
 public class ChatRoomMembersPK implements Serializable {
 
-    private int chat_room_id;
-    private int member_id;
+    private Long chat_room_id;
+    private Long member_id;
 
     // 기본 생성자, equals(), hashCode() 메소드 구현
 }
